后台发送前端消息 id改为byte

+++ client/src/main/scala/com/neo/sk/gypsy/actor/BotActor.scala
@@ -189,7 +189,7 @@ object BotActor {
       msg match {
         case Action(key,swing,sender) =>
           botHolder.gameActionReceiver(key,swing)
-          sender ! ActionRsp(frameIndex = botHolder.getFrameCount.toInt, msg = "ok")
+          sender ! ActionRsp(frameIndex = botHolder.getFrameCount, msg = "ok")
           Behaviors.same
